+++ neqo-common/src/lib.rs
@@ -11,7 +11,6 @@ mod codec;
 mod datagram;
 mod incrdecoder;
 pub mod log;
-pub mod once;
 pub mod timer;
 
 pub use self::codec::{Decoder, Encoder};
