+++ neqo-crypto/src/once.rs
